Pull out a core module from the main implementtaion
